requirements_dev.txt
deep-translator==1.5.4
fastapi==0.68.1
flake8==3.9.2
mysql-connector-python==8.0.26
pandas==1.3.3
pipdeptree==2.2.0
python-dotenv==0.19.0
uvicorn==0.15.0
beautifulsoup4==4.10.0
tqdm==4.62.3
ncbi-datasets-pylib==13.8.0
isort==5.10.1
black==22.6.0
pylint==2.14.4
tabulate==0.9.0
fastapi-cache2==0.1.8
